split up AliasTracker into a separate file
