+++ torch/csrc/jit/passes/alias_analysis.cpp
@@ -2,7 +2,6 @@
 
 #include <torch/csrc/jit/script/error_report.h>
 #include <torch/csrc/utils/memory.h>
-#include <queue>
 
 namespace torch {
 namespace jit {
@@ -22,375 +21,6 @@ bool shouldAnnotate(const Value* v) {
   return shouldAnnotate(v->type());
 }
 } // namespace
-
-// class AliasTracker
-//
-// This class tracks the "A points to B" graph for all values, as well as
-// wildcards and writes. It is used by AliasDb to provide a higher-level API.
-//
-// NOTE: this implementation is not very efficient; it's designed to be easy to
-// mutate as you modify the graph.
-class AliasTracker {
- public:
-  // Returns true iff `v` is present in the alias set tracker.
-  bool contains(const Value* v) const {
-    return map_.count(v);
-  }
-
-  bool writesTo(Node* n, const Value* v) const {
-    if (isWildcard(v)) {
-      return wildcardWriters_.count(n);
-    }
-
-    if (!map_.count(v)) {
-      return false;
-    }
-
-    return map_.at(v)->writers.count(n);
-  }
-
-  // Whether `a` *may* point to `b`
-  bool pointsTo(const Value* a, const Value* b) const {
-    if (!map_.count(a)) {
-      return false;
-    }
-    if (isWildcard(a) || isWildcard(b)) {
-      return true;
-    }
-
-    // BFS the subtree where the root is `a`s element and the branches are the
-    // `pointsTo` relationships.
-    const auto root = map_.at(a);
-    return root->bfs(
-        [&](const Element* el) { return el->value == b; },
-        BfsDirection::POINTS_TO,
-        /*shortCircuit=*/true);
-  }
-
-  // Make `v` point at `to`.
-  void makePointerTo(const Value* v, const Value* to) {
-    if (v == to) {
-      return;
-    }
-
-    // If `to` is a wildcard, don't insert anything into the graph; wildcards
-    // are tracked separately since they have different aliasing rules.
-    if (isWildcard(to)) {
-      setWildcard(v);
-      return;
-    }
-
-    if (!map_.count(to)) {
-      makeFreshValue(to);
-    }
-
-    if (!map_.count(v)) {
-      makeFreshValue(v);
-    }
-
-    auto vEl = map_.at(v);
-    auto toEl = map_.at(to);
-
-    vEl->pointsTo.insert(toEl);
-    toEl->pointedFrom.insert(vEl);
-  }
-
-  // Give `v` a fresh alias (i.e. it does not point to any value)
-  void makeFreshValue(const Value* v) {
-    auto el = torch::make_unique<Element>();
-    el->value = v;
-
-    auto rawPtr = el.get();
-    elements_.emplace(rawPtr, std::move(el));
-    map_.emplace(v, rawPtr);
-  }
-
-  // Register `v` as a wildcard value.
-  void setWildcard(const Value* v) {
-    wildcards_.insert(v);
-  }
-
-  // is `v` a wildcard?
-  bool isWildcard(const Value* v) const {
-    return wildcards_.count(v);
-  }
-
-  // Register the fact that `n` writes to `v`.
-  void registerWrite(const Value* v, Node* n) {
-    numWrites_++;
-
-    if (isWildcard(v)) {
-      wildcardWriters_.insert(n);
-      return;
-    }
-
-    AT_ASSERT(map_.count(v));
-    map_.at(v)->writers.insert(n);
-  }
-
-  // Return all aliases of `v`. This is the full set of any other value that
-  // *may* represent the same memory location.
-  // NOTE: this does not consider wildcard values
-  std::unordered_set<const Value*> getAliases(const Value* v) const {
-    std::unordered_set<const Value*> ret;
-    if (!map_.count(v)) {
-      return ret;
-    }
-
-    const auto root = map_.at(v);
-
-    root->bfs(
-        [&](const Element* el) {
-          ret.insert(el->value);
-          return false; // fn has to return bool but we don't use the result
-        },
-        BfsDirection::BOTH);
-    return ret;
-  }
-
-  // Get all nodes that write to `v` or a value that may alias `v`.
-  std::unordered_set<Node*> getWrites(const Value* v) const {
-    std::unordered_set<Node*> ret;
-    if (!map_.count(v)) {
-      return ret;
-    }
-
-    // Any write to a wilcard may write to `v`.
-    for (auto writer : wildcardWriters_) {
-      ret.insert(writer);
-    }
-
-    if (useCache_) {
-      for (auto writer : getWritersCached(v)) {
-        ret.insert(writer);
-      }
-      return ret;
-    }
-
-    const auto root = map_.at(v);
-    root->bfs(
-        [&](const Element* el) {
-          for (auto writer : el->writers) {
-            ret.insert(writer);
-          }
-          return false; // fn has to return bool but we don't use the result
-        },
-        BfsDirection::BOTH);
-
-    return ret;
-  }
-
-  // Functionally equivalent to getWrites().size() > 0, but with a
-  // short-circuiting implementation to be faster.
-  bool hasWriters(const Value* v) const {
-    if (!map_.count(v)) {
-      return false;
-    }
-
-    if (isWildcard(v)) {
-      // If `n` has a wildcard, any write in the graph may write to it.
-      // So the only way we know there are no writers is if there are no writes
-      // at all.
-      return numWrites_ == 0;
-    }
-
-    if (wildcardWriters_.size() > 0) {
-      // A write to the wildcard may be a write to any value.
-      return true;
-    }
-
-    if (useCache_) {
-      return hasWritersCached(v);
-    }
-
-    const auto root = map_.at(v);
-    return root->bfs(
-        [&](const Element* el) { return el->writers.size() > 0; },
-        BfsDirection::BOTH,
-        /*shortCircuit=*/true);
-  }
-
-  // Get all nodes that write to a wildcard value.
-  const std::unordered_set<Node*>& getWildcardWriters() const {
-    return wildcardWriters_;
-  }
-
-  void dump() const {
-    std::cout << "\n===2. ALIAS DB===\n";
-    for (const auto& ptrPair : elements_) {
-      const auto element = ptrPair.first;
-      if (element->pointsTo.size() > 0) {
-        std::cout << element->value->uniqueName() << " points to: ";
-        for (const auto pointedTo : element->pointsTo) {
-          std::cout << pointedTo->value->uniqueName() << ", ";
-        }
-        std::cout << "\n";
-      }
-    }
-
-    std::cout << "\n===3. WILDCARDS===\n";
-    for (const auto wildcard : wildcards_) {
-      std::cout << wildcard->uniqueName() << ", ";
-    }
-    std::cout << "\n";
-  }
-
- private:
-  enum class BfsDirection {
-    POINTS_TO,
-    POINTED_FROM,
-    // Consider both pointer directions. The closure obtained from this
-    // represents the whole "alias set" of a value.
-    BOTH
-  };
-  // `Element` represents the vertex in the points-to graph. It has a 1:1
-  // relationship with IR `Value`s.
-  struct Element {
-    const Value* value = nullptr;
-    // All values that this value *may* point to. It's possible to have multiple
-    // values that you might point to due to control flow/complex ops
-    std::unordered_set<Element*> pointsTo;
-    // Backreference to values that point to `this`
-    std::unordered_set<Element*> pointedFrom;
-    // Nodes that write to this specific value.
-    std::unordered_set<Node*> writers;
-
-    // Do a breadth-first search over the graph, starting at `this` and
-    // traversing in the direction `dir`.`fn` will be run on each element.
-    //
-    // If `shortCircuit` is set, then if `fn` evaluates to true the search will
-    // short-circuit and return true. You can use this to do existence checks
-    // on the graph or whatever.
-    template <typename Fn>
-    bool bfs(Fn fn, BfsDirection dir, bool shortCircuit = false) const {
-      std::queue<const Element*> queue;
-      std::unordered_set<const Element*> seen;
-
-      queue.push(this);
-      while (!queue.empty()) {
-        const auto el = queue.front();
-        queue.pop();
-        seen.insert(el);
-
-        if (fn(el) && shortCircuit) {
-          return true;
-        }
-
-        switch (dir) {
-          case BfsDirection::POINTS_TO: {
-            for (auto ptr : el->pointsTo) {
-              if (!seen.count(ptr)) {
-                queue.push(ptr);
-              }
-            }
-          } break;
-
-          case BfsDirection::POINTED_FROM: {
-            for (auto ptr : el->pointedFrom) {
-              if (!seen.count(ptr)) {
-                queue.push(ptr);
-              }
-            }
-          } break;
-
-          case BfsDirection::BOTH: {
-            for (auto ptr : el->pointsTo) {
-              if (!seen.count(ptr)) {
-                queue.push(ptr);
-              }
-            }
-            for (auto ptr : el->pointedFrom) {
-              if (!seen.count(ptr)) {
-                queue.push(ptr);
-              }
-            }
-          } break;
-        }
-      }
-      return false;
-    }
-  };
-
-  // Structure that owns all the element pointers. It's a map of
-  //  raw pointer -> unique_ptr to facilitate easy queries
-  std::unordered_map<Element*, std::unique_ptr<Element>> elements_;
-  // Index to look up whatever element corresponds to that value.
-  std::unordered_map<const Value*, Element*> map_;
-  // All values that may point to a wildcard value.
-  std::unordered_set<const Value*> wildcards_;
-  // All nodes that write to a wildcard
-  std::unordered_set<Node*> wildcardWriters_;
-  size_t numWrites_ = 0;
-
-  /**
-   * Caching layer.
-   */
-  using set_id_t = size_t;
-  bool useCache_ = true;
-  mutable std::unordered_map<const Element*, std::unordered_set<set_id_t>>
-      elementToSet_;
-  mutable std::unordered_map<set_id_t, std::unordered_set<Node*>> setToWrites_;
-  mutable bool cacheStale_ = true;
-  mutable set_id_t lastId = 0;
-
-  // Cache results in a way to make common queries constant time.
-  void cache() const {
-    if (!cacheStale_) {
-      return;
-    }
-
-    for (const auto& pr : elements_) {
-      const auto el = pr.first;
-      // For each value that does point to anything, assign a fresh set.
-      if (el->pointsTo.size() == 0) {
-        const auto id = getFreshId();
-        assignSet(el, id);
-
-        // Propagate this set to every element that points to `el`
-        el->bfs(
-            [&](const Element* pointerTo) { return assignSet(pointerTo, id); },
-            BfsDirection::POINTED_FROM);
-      }
-    }
-
-    cacheStale_ = false;
-  }
-
-  bool hasWritersCached(const Value* v) const {
-    cache();
-    for (const auto& set : elementToSet_.at(map_.at(v))) {
-      if (setToWrites_.count(set) && setToWrites_.at(set).size() > 0) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  std::unordered_set<Node*> getWritersCached(const Value* v) const {
-    cache();
-    std::unordered_set<Node*> ret;
-    for (const auto& set : elementToSet_.at(map_.at(v))) {
-      if (setToWrites_.count(set) > 0) {
-        for (auto write : setToWrites_.at(set)) {
-          ret.insert(write);
-        }
-      }
-    }
-    return ret;
-  }
-
-  bool assignSet(const Element* el, set_id_t id) const {
-    elementToSet_[el].insert(id);
-    for (auto write : el->writers) {
-      setToWrites_[id].insert(write);
-    }
-    return true;
-  }
-
-  set_id_t getFreshId() const {
-    return ++lastId;
-  };
-};
 
 AliasDb::~AliasDb() = default;
